+++ packages/isar_generator/lib/src/code_gen/collection_schema_generator.dart
@@ -75,6 +75,9 @@ String _generateSchema(ObjectInfo object) {
         {
           'name': property.isarName,
           'type': property.isarType.name,
+          if (property.isarType == IsarType.object ||
+              property.isarType == IsarType.objectList)
+            'target': property.scalarDartType,
         },
     ],
     'indexes': [