+++ packages/isar_generator/lib/src/isar_type.dart
@@ -14,6 +14,7 @@ enum IsarType {
   long,
   double,
   dateTime,
+  enumeration,
   string,
   object,
   boolList,
@@ -23,6 +24,7 @@ enum IsarType {
   longList,
   doubleList,
   dateTimeList,
+  enumerationList,
   stringList,
   objectList,
 }
@@ -36,16 +38,6 @@ extension IsarTypeX on IsarType {
     return index > IsarType.string.index;
   }
 
-  bool get containsFloat {
-    return this == IsarType.float ||
-        this == IsarType.double ||
-        this == IsarType.floatList ||
-        this == IsarType.doubleList;
-  }
-
-  bool get containsString =>
-      index == IsarType.string.index || index == IsarType.stringList.index;
-
   int get staticSize {
     if (this == IsarType.bool || this == IsarType.byte) {
       return 1;
@@ -56,76 +48,6 @@ extension IsarTypeX on IsarType {
     }
   }
 
-  int get elementSize {
-    switch (this) {
-      case IsarType.boolList:
-      case IsarType.byteList:
-        return 1;
-      case IsarType.intList:
-      case IsarType.floatList:
-        return 4;
-      case IsarType.longList:
-      case IsarType.doubleList:
-      case IsarType.dateTimeList:
-        return 8;
-      // ignore: no_default_cases
-      default:
-        return 0;
-    }
-  }
-
-  IsarType get scalarType {
-    switch (this) {
-      case IsarType.boolList:
-        return IsarType.bool;
-      case IsarType.byteList:
-        return IsarType.byte;
-      case IsarType.intList:
-        return IsarType.int;
-      case IsarType.floatList:
-        return IsarType.float;
-      case IsarType.longList:
-        return IsarType.long;
-      case IsarType.doubleList:
-        return IsarType.double;
-      case IsarType.dateTimeList:
-        return IsarType.dateTime;
-      case IsarType.stringList:
-        return IsarType.string;
-      case IsarType.objectList:
-        return IsarType.object;
-      // ignore: no_default_cases
-      default:
-        return this;
-    }
-  }
-
-  IsarType get listType {
-    switch (this) {
-      case IsarType.bool:
-        return IsarType.boolList;
-      case IsarType.byte:
-        return IsarType.byteList;
-      case IsarType.int:
-        return IsarType.intList;
-      case IsarType.float:
-        return IsarType.floatList;
-      case IsarType.long:
-        return IsarType.longList;
-      case IsarType.double:
-        return IsarType.doubleList;
-      case IsarType.dateTime:
-        return IsarType.dateTimeList;
-      case IsarType.string:
-        return IsarType.stringList;
-      case IsarType.object:
-        return IsarType.objectList;
-      // ignore: no_default_cases
-      default:
-        return this;
-    }
-  }
-
   String get name {
     switch (this) {
       case IsarType.id:
@@ -133,6 +55,7 @@ extension IsarTypeX on IsarType {
       case IsarType.bool:
         return 'Bool';
       case IsarType.byte:
+      case IsarType.enumeration:
         return 'Byte';
       case IsarType.int:
         return 'Int';
@@ -150,6 +73,7 @@ extension IsarTypeX on IsarType {
       case IsarType.boolList:
         return 'BoolList';
       case IsarType.byteList:
+      case IsarType.enumerationList:
         return 'ByteList';
       case IsarType.intList:
         return 'IntList';
@@ -165,43 +89,6 @@ extension IsarTypeX on IsarType {
       case IsarType.objectList:
         return 'ObjectList';
     }
-  }
-
-  String dartType(bool nullable, bool elementNullable) {
-    final nQ = nullable ? '?' : '';
-    final nEQ = elementNullable ? '?' : '';
-    switch (this) {
-      case IsarType.id:
-        return 'Id$nQ';
-      case IsarType.bool:
-        return 'bool$nQ';
-      case IsarType.byte:
-      case IsarType.int:
-      case IsarType.long:
-        return 'int$nQ';
-      case IsarType.double:
-      case IsarType.float:
-        return 'double$nQ';
-      case IsarType.dateTime:
-        return 'DateTime$nQ';
-      case IsarType.string:
-        return 'String$nQ';
-      case IsarType.boolList:
-        return 'List<bool$nEQ>$nQ';
-      case IsarType.byteList:
-        return 'Uint8List$nQ';
-      case IsarType.intList:
-      case IsarType.longList:
-        return 'List<int$nEQ>$nQ';
-      case IsarType.floatList:
-      case IsarType.doubleList:
-        return 'List<double$nEQ>$nQ';
-      case IsarType.dateTimeList:
-        return 'List<DateTime$nEQ>$nQ';
-      case IsarType.stringList:
-        return 'List<String$nEQ>$nQ';
-    }
-    throw '';
   }
 }
 
@@ -234,6 +121,8 @@ IsarType? _getPrimitiveIsarType(DartType type) {
     return IsarType.string;
   } else if (_isDateTime(type.element!)) {
     return IsarType.dateTime;
+  } else if (type.isDartCoreEnum) {
+    return IsarType.enumeration;
   }
 
   return null;
@@ -251,12 +140,32 @@ IsarType? getIsarType(DartType type, Element element) {
     final parameterizedType = type as ParameterizedType;
     final typeArguments = parameterizedType.typeArguments;
     if (typeArguments.isNotEmpty) {
-      final listType = typeArguments[0];
-      final primitiveType = _getPrimitiveIsarType(listType);
-      if (primitiveType == IsarType.id) {
-        err('Id lists are not supported.', element);
-      } else {
-        return primitiveType?.listType;
+      switch (_getPrimitiveIsarType(typeArguments[0])) {
+        case IsarType.id:
+          err('Id lists are not supported.', element);
+        case IsarType.bool:
+          return IsarType.boolList;
+        case IsarType.byte:
+          return IsarType.byteList;
+        case IsarType.int:
+          return IsarType.intList;
+        case IsarType.float:
+          return IsarType.floatList;
+        case IsarType.long:
+          return IsarType.longList;
+        case IsarType.double:
+          return IsarType.doubleList;
+        case IsarType.dateTime:
+          return IsarType.dateTimeList;
+        case IsarType.enumeration:
+          return IsarType.enumerationList;
+        case IsarType.string:
+          return IsarType.stringList;
+        case IsarType.object:
+          return IsarType.objectList;
+        // ignore: no_default_cases
+        default:
+          throw UnimplementedError();
       }
     }
   }