+++ packages/isar_test/test/mutli_type_model.dart
@@ -36,10 +36,6 @@ class MultiTypeModel {
 
   String stringValue = '';
 
-  Uint8List? bytesValueN;
-
-  Uint8List bytesValue = Uint8List(0);
-
   List<bool> boolList = [];
 
   List<bool?> boolNList = [];
@@ -47,6 +43,14 @@ class MultiTypeModel {
   List<bool>? boolListN;
 
   List<bool?>? boolNListN;
+
+  Uint8List? byteListN;
+
+  List<byte>? byteList2N;
+
+  Uint8List byteList = Uint8List(0);
+
+  List<byte> byteList2 = [];
 
   List<short> intList = [];